backend/requirements.txt
discord.py
google-generativeai
python-dotenv
pillow
supabase